refactor: Rename `…/events/websockets/lib/pickWebsocketsTemplatePart.js`
