+++ lib/plugins/aws/package/compile/events/websockets/lib/deployment.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const crypto = require('crypto');
-const pickWebsocketsTemplatePart = require('./pickWebsocketsTemplatePart');
+const pickWebsocketsTemplatePart = require('./pick-websockets-template-part');
 
 module.exports = {
   compileDeployment() {
