+++ lib/plugins/aws/package/compile/events/websockets/lib/pick-websockets-template-part.js
